Delete legacy profile achievements
