+++ Stepic/Legacy/Controllers/Profile/Achievements/AchievementProgressData.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+struct AchievementProgressData {
+    let currentScore: Int
+    let maxScore: Int
+    let currentLevel: Int
+    let maxLevel: Int
+    let kind: String
+}
