--- Stepic/Legacy/Controllers/Profile/Achievements/ProfileAchievementsContentView.xib
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="15705" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="15706"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="ProfileAchievementsContentView" customModule="Stepic" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="375" height="98"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <stackView opaque="NO" contentMode="scaleToFill" distribution="fillEqually" spacing="8" translatesAutoresizingMaskIntoConstraints="NO" id="736-Rh-JdR">
-                    <rect key="frame" x="24" y="10" width="327" height="80"/>
-                    <constraints>
-                        <constraint firstAttribute="height" priority="999" constant="80" id="XxQ-Wx-SLf"/>
-                    </constraints>
-                </stackView>
-                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="WOt-0i-A7a">
-                    <rect key="frame" x="163.5" y="27" width="48" height="44"/>
-                    <constraints>
-                        <constraint firstAttribute="height" constant="44" id="6Lj-5w-T9C"/>
-                    </constraints>
-                    <fontDescription key="fontDescription" type="system" weight="light" pointSize="16"/>
-                    <state key="normal" title="Button"/>
-                    <connections>
-                        <action selector="onRefreshButtonClick:" destination="iN0-l3-epB" eventType="touchUpInside" id="oIX-7w-CHU"/>
-                    </connections>
-                </button>
-            </subviews>
-            <constraints>
-                <constraint firstItem="WOt-0i-A7a" firstAttribute="centerY" secondItem="iN0-l3-epB" secondAttribute="centerY" id="06z-3d-69B"/>
-                <constraint firstItem="736-Rh-JdR" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="24" id="2oa-wX-37n"/>
-                <constraint firstItem="WOt-0i-A7a" firstAttribute="centerX" secondItem="iN0-l3-epB" secondAttribute="centerX" id="KrM-3N-YSd"/>
-                <constraint firstAttribute="bottom" secondItem="736-Rh-JdR" secondAttribute="bottom" priority="999" constant="8" id="XkK-7d-hdg"/>
-                <constraint firstItem="736-Rh-JdR" firstAttribute="top" secondItem="vUN-kp-3ea" secondAttribute="top" priority="999" constant="10" id="lgy-7A-Zh8"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="736-Rh-JdR" secondAttribute="trailing" constant="24" id="xUN-Aw-y6h"/>
-            </constraints>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
-            <connections>
-                <outlet property="achievementsStackView" destination="736-Rh-JdR" id="0bm-NF-hFL"/>
-                <outlet property="refreshButton" destination="WOt-0i-A7a" id="ZEv-AK-VYH"/>
-            </connections>
-            <point key="canvasLocation" x="24.5" y="-239.5"/>
-        </view>
-    </objects>
-</document>
